Done with fixes, tested on mock project

+++ assetMaker.py
@@ -17,10 +17,9 @@ def makeExtension():
 
 	userProjects = listProjectsUsingAsset()
 
-	# updateAll = promptPushToAll()
+	updateAll = promptPushToAll()
 	for project in userProjects:
-		if True:
-		# if (updateAll or promptPushToProject(project)):
+		if (updateAll or promptPushToProject(project)):
 			deleteResourcesFromProject(project)
 			excludeAssetResourcesFromProject(project)
 			copyResourcesToProject(project)
